Update README.md
A task manager, where the user can add tasks and mark them as done. The program also allows the user to reset all tasks when necessary.
Project was developed in Python, using the PySimpleGUI library. The objective of this project was the implementation of a graphical interface using the python language.